refactor: move tests into subfolders
This will make it easier to store tutorials and other documentation artifacts close to the tests they represent.
