+++ packages/tests-postman-wip/lib/index.mjs
@@ -1,0 +1,12 @@
+/* eslint-disable import/extensions */
+import TestAccessToken from './access-token/index.mjs';
+import TestDidConfiguration from './did-configuration/index.mjs';
+import TestCredentialsIssue from './credentials-issue/index.mjs';
+import TestPresentationsProve from './presentations-prove/index.mjs';
+
+export default {
+  TestAccessToken,
+  TestDidConfiguration,
+  TestCredentialsIssue,
+  TestPresentationsProve,
+};
